+++ libs/ui/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @dhi/arsenal.ui
 
+## 14.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @dhi/arsenal.models@4.0.0
+
 ## 13.0.0
 
 ### Minor Changes